requirements.txt
# Core async scraping
aiohttp>=3.8.0
lxml>=4.9.0

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0